remove commit_verification_keys
The latest version of the github resource doesn't need this (since
alphagov/gsp#378).  Instead the process is to add your GPG key to
GitHub and get GitHub to verify the signature; we only need to
validate the signature on the GitHub merge commit against the GitHub
web flow key.

Note that if we can get rid of commit_verification_keys from all our
pipelines, we can also merge alphagov/gds-trusted-developers#111 so
that we no longer have to faff with GPG keys in that repository when
people join teams that commit to GSP projects.

This also bumps the task-toolbox and github-resource references to use
the tag provided by GSP in the `concourse` secret.