EID-1762 Remove `@NotNull` annotations from `TranslatedHubResponse`
The TranslatedHubResponse is created from a response from the VSP. If
the VSP encounters a SAML Response with a status of `REQUESTER`,
`NO_AUTHN_CONTEXT`, or `AUTHN_FAILED` it returns a
`TranslatedNonMatchingResponseBody` with pid, loa, and attributes set to
`null`. These annotations are being removed to reflect that.

As an aside, these annotations don't actually seem to validate an object
created from a Response, which is why this hasn't caused an issue
before.