EID-1822: Use the Amazon Corretto image in all the places
This image is provided by Amazon and has JDK 11 installed with some Java patches.
The logic to install Java has been removed from the VSP as the new image comes with Java pre-installed.
Build apps and run tests in the pipeline also using the Corretto image instead of the vanilla openjdk11.

Author: @CharlesIC