WIP
Skeleton app that can consume a particular config file format and regurgitate that config via a couple of endpoints that produce disgraceful html (if you can call it that).

Also, unmarshals a base64 string in the config straight into a KeyStore which can be used straight away.

First attempt at *-deployment.yaml and *-service.yaml which are currently copy and pastes of the esp versions with esp replaced with metadata-service.