EID-1848 Deploy countries configMap
This adds tasks to the build pipeline to create a configMap for the
consumption by the proxy-node's metadata service, containing the
configuration for the countries.

The config is defined in the `verify-eidas-config` repo, as well as the
tooling to generate the truststore required. The resulting config file
from running the tool is used to populate a configMap.

We use kubectl with the `--dry-run` flag to generate the yaml resource
definition. We do this as we also use the `--from-file` flag - it's not
possible to define a configMap with static kubeyaml that references a
file in this way.

We edit the output of kubectl with yq to add a couple of annotations.
These annotations are used by kapp. They allow kapp to version the
configMap. This is required so that when we deploy the metadata-service
it is aware of which specific version of the configMap it using. This way,
when the configMap changes and is redeployed, kapp knows that
metadata-service needs to also be redeployed - to pick up the latest
version of the configMap. The docs for this versioning [can be found
here](https://github.com/k14s/kapp/blob/master/docs/faq.md#updating-deployments-when-configmap-changes)

The versioning is only necessary as the metadata service currently only
reads the config file (mounted from the configMap) on startup. If it
could be designed to automatically "hot-reload" from the file we could
skip this, as the mounted config file is automatically updated when the
configMap is updated.

The final yaml resource definition for the configMap is added to the
`manifests` directory, which is where the other manifests are placed by
the `helm template` command. This allows the existing call to `kapp` to
deploy the configMap along with the apps themselves.

Co-authored-by: John Watts <john.watts@digital.cabinet-office.gov.uk>