EID-1826 Create skeleton dropwizard app
It will be responsible for fetching metadata and serving it to the
proxy-node.

Skeleton drop wizard app to consume config provided by the pipeline. The
theory is that having something at either end of the pipe will make it
easier to join things up.

Skeleton app that can consume a particular config file format and
regurgitate that config via a couple of endpoints that produce
disgraceful html (if you can call it that).

Also, unmarshals a base64 string in the config straight into a KeyStore
which can be used straight away.

Co-authored-by: John Watts <john.watts@digital.cabinet-office.gov.uk>