BAU: Upgrade test libraries and Gradle to the latest versions
- Use the Gradle `implementation` and `api` dependency configurations instead of `compile` as this is deprecated
`implementation` replaces `compile` for dependencies that don't need to be access by users of a library. `api` exposes a dependency to the users. The old `compile` configuration - which is equivalent to `api` - used to expose all the compile-time dependencies to the callers. This is now phased out in favour of hiding compile-time dependencies by default and explicitly specifying which libraries should be exposed. More info on this in [Gradle Java Library plugin docs](https://docs.gradle.org/current/userguide/java_library_plugin.html#sec:java_library_separation)

- Remove the pom generation task
The `maven` plugin we used to generate the poms will be retired in the next version of Gradle. This was intended to be a workaround for Snyk GitHub integration not supporting Gradle projects but it wasn't a particularly good solution anyway. Tt will ba abandoned in favour of using new Gradle features introduced in Gradle 6. This hasn't been tested yet but Gradle now supports publishing more detailed metadata about modules and dependencies. We might be able to use that or wait until the Snyk GitHub integration introduces support for Gradle. We also have the manual Snyk script so we're covered for vulnerability detection.

- Apply new way of configuring the Gradle build scan plugin

- Stop using deprecated assertions and methods
- Use generics in `Attributes` to remove unchecked casting of list types
- Other minor refactorings and code tidy-up

Author: @CharlesIC