EID-1854 Smoke Test Infrastructure
The idea of this PR is to reuse the existing image we build for acceptance tests to also run smoke tests.
Therefore in our build pipeline we target only the acceptance test cucumber feature, plus introduce a new smoke test cucumber feature, which will be run in the new smoke-test pipeline.

This idea is being tested out on the release branch, as it has an integration release.