EID-1876 Hub error page if can't build error saml
When there is an error in response from translator to gateway, for a subset of errors we'll try to build an error saml response, by calling translator again to generate a signed error saml response.

This is handled by the translator `/generate-failure-response` resource.

We do not handle the case where this resource throws an ApplicationException - and a standard dropwizard error page is rendered.

This PR is to redirect to the hub error page in this scenario.

https://govukverify.atlassian.net/browse/EID-1876