EID-1873 delete integration test connector on deploy
To reduce the creation of handles to CloudHSM keys, we redeploy our integration instances daily.
The current daily roll is however not deleting the connector pod, CloudHSM handles are not decreasing.

This commit explicity deletes the connector pod in the current namespace during a deploy.
It's a bit of a band-aid, but is required to mitigate incidents during December holidays.

NB. This change has been proven to work in the sandbox cluster.

https://govukverify.atlassian.net/browse/EID-1873