EID-1877 Decouple Stub Connector from the HSM
TLDR; The Stub Connector represents a stub country. It should not have its saml signing coupled to an HSM key.

The Stub Connector currently has its metadata certs+keys generated via the verify-metadata-controller, and the saml signing key is held in the HSM, see https://test-integration-connector.london.verify.govsvc.uk/ConnectorMetadata

Every SAML request from stub connector is currently signed by a key in the CloudHSM. This was done as a convenience as the truststore for this signing is also generated via verify-metadata-controller, and easily made available to the ESP (the service that reads the connector metadata).

We want to model a a stub country that has its own PKI completely outside the HSM. The eventual goal is to run the Stub Connector in PaaS, to complete the separation.

This commit will:

* Create a pipeline to set a metadata signing cert+key, saml signing cert+key plus a metadata signing truststore, all from verify-dev-pki, into a cluster-wide configmap named test-pki-configmap
* Read test-pki-configmap to set a truststore in ESP
* Read test-pki-configmap via a volumeMount to provide metadata and saml signing credentials to the connector
* Remove HSM client config from stub connector
* Publish connector node metadata from a Resource using a mustache template. The method is equivalent to how connector node metadata is currently created in the verify-metadata-controller
* Use sensible default values in connector config.yml to avoid setting config via env vars