EID-1827 Fetch countries' metadata
Use verify-saml-libs 232 and apply the changes in https://github.com/alphagov/verify-proxy-node/pull/412

Tidy up unused imports.

Replace some sysouts with ProxyNodeLogger

MetadataResolverService to retrieve and cache metadata from the URLs provided in EidasConfig.

Add "metadata" endpoint as temporary endpoint for use in tests.  End point will currently return a reponse with status 200 and the metadata entity id for successfully resolved metadata and 500 if there are any problems.  This will be replaced with the endpoint returning the MetatronResponse [sic] DTO and the tests will be updated accordingly in future.

Provide initial test classes.  The CreatesConfigDropwizardClientRule writes the countriesConfig.yaml used by MetatronAppRule to /tmp and makes the temporary port that it will listen on available.

TestMetadataResource uses the getPort method of CreatesConfigDropwizardClientRule in setting the entityId of the metadata resources it creates.

MetatronTests tests that the responses are as expected for the various metadatas gathered by Metatron(AppRule).  When the correct test-metadatas are provided to TestMetadataResource then metatron will successfully resolve two of them and fail to resolve / provide metadata that is, unsigned, expired or untrusted.