EID-1887 Add MetatronResponse DTO and validation tests
The tests included here, as well as the interface and DTO with
annotations are purely a learning exercise but worth keeping in the git
history for posterity. As such, i'll be removing them in the next
commit.

The tests are to show that adding validations such as `@NotNull` to a
DTO doesn't actually do anything if you are receiving the DTO as a
response. The validations only get run if the DTO is used to generate an
object from JSON POSTed to an endpoint you control.

You _can_ manually run validations on DTO's you receive as a response,
by creating a validator and using it on the object. [See here for an
example](https://www.baeldung.com/javax-validation#programmatic)