EID-1828 Create Metatron API endpoint
https://govukverify.atlassian.net/browse/EID-1828

Provide a new Rest API which will be called from the ESP And Translator.
Given a country's connector node entity id, the API provides data about that country and its metadata.

Unit tests have been dropped in favour of testing real cases with Dropwizard AppRule tests.

The in-memory map of country -> metadata resolver in MetadataResolver service now uses EidasCountryConfig as key.
This is so we can store country info in the key, like country code.

The API returns either a:
* 200 with a json representation of CountryMetadataResponse
* 400 with a json error map when the client is sending a bad request
* 500 with a json error map when the metatron has a processing error