BAU: Fix docker local-startup and add Metatron
The way that stub-connector got it's signing credentials had changed
slightly which prevented it from starting up.

This also adds Metatron although none of the apps are currently
configured to use it. This can change once the work has been completed
for them to start using it.

Also bumps the gradle version the images are built on to match the new
gradle syntax being used.