EID-1836: Use Metatron proxy and values in translator
This PR uses the shared Metatron proxy to fetch values from the countries
metadata and use the encryption certificate and country code when
generating and eIDAS response.

There is a slight change to the gateways session object to allow the
entity ID to be passed through to the translator - this is necessary so
that the translator can call Metatron correctly.

As the translator is now fetching the metadata for each requests we can
remove the hard coded values for each country from the configuration
class and yaml.

Co-authored-by: Olakunle Jegede <olakunle.jegede@digital.cabinet-office.gov.uk>
Co-authored-by: Phil Miller <phillip.miller@digital.cabinet-office.gov.uk>