EID-1836: Remove connector encryption cert from session
The connector node encryption certificate was previously being passed
from the ESP, through the gateway (via the session storage) and on to
the translator.

The translator is now fetching the certificate from Metatron directly,
so we don't need to pass it through. This means we can remove it from
the DTO as well as from the session.

This commit also renames some fields to make it clear that when we refer
to an issuer, we're actually talking about their entityId.