BAU Un-break the broken build
Some time ago, builds stopped building.  The test that fails doesn't fail locally or when Travis runs it.  This suggests some kind of race is going on and the wrong thing wins in the pipeline.  It looks like it's the mockatron that's providing the null pointer exceptions when it should be providing port numbers so I've changed the order in which the RuleChain is executed to see if that helps.