EID-1917: Add smoke tests to all countries connected to Connector Node
The tests all start from the 'Personal tax account' service, from there sign in with a digital identity from another European country is selected. Then the country of choice and then we check that we have reached the countries hub by asserting we can see specialised text on the page.