EID-1894: Add additional yaml files
Copied principal.yaml and image-repositories.yaml from the ci/build directory. It creates an ECR Image Repository for your docker images to be pushed to, along with some credentials for it.

Copied artefact-signing-key-secret.yaml from ci/prod /integration. The artefact signing key is used to gpg sign the release that gets pushed to github