EID-1964: Add alternative DNS names for the ingress TLS certs that are short enough for CN
Since we had to add the namespace to the URLs, the default config tried to generate Let's Encrypt certs with the new URLs as the DNS name and by default the CN (Common Name). However, the CN has a limit of 64 chars and the long URLs exceeded that. As as solution, add a DNS name to the list that's short enough to be picked as a CN for the new certs.

Author: @CharlesIC