Produce metrics on presence of country metadata
Create a scheduled runnable in metatron that retrieves cached country metadata.
Report 1 or 0 to a Prometheus Gauge depending on presence of metadata.

Country metadata is stored in stateful MetadataResolver object, which retrieves metadata at intervals, and caches the result.
If there is no cached metadata, a 0 will be sent to the Gauge.

The scheduler runs the check every 5 minutes.