EID-1911: Set the transient field in the translator
If the pid is transient then we append _tr to the pid
Add additional tests for NameId scenarios
Added tests to show the result of transientPid if the NameId is UNSPECIFIED, PERSISTANT or null
Created a SecureRandomIdentifierGenerationStrategy to create the new pid if the NameId is transient
Set the field on construction & create a method to assign the transient pid. By doing this it makes it easier for testing
Also we don't evaluate NameIDPolicy until this point, so a request could be crafted without a NameID and we would throw a NPE here.
A null check should be done here, and when null, transientPid should be false