BAU: Misc Stub Connector improvements
  - Show additional information on the Start Connector Page

Display links to Proxy Node metadata and certs, and the Stub Connector metadata from Metatron. This allows access to the most common things we need to look at when we do testing. Having the link to the Metatron metadata allows us to check that the Metatron has initialised correctly and has fetched SC metadata before running a journey to avoid errors.

Substitute "host.docker.internal" for "localhost" when running locally in Docker. Construct the certs URL by removing the last "/" and appending the path. This will also work in build and integration.

  - Remove explicit Proxy Node metadata health check

We already get this for free from the Metadata Bundle that we use to resolve PN metadata

  - Misc fixes to spacing