BAU: Improvements to local running
  - Upgrade gradle to 6.4.1
  - Give local gateway metadata certs
  - Add shutdown script
  - Run installDist task by default
  - Give containers nice names
  - Open Stub Connector page in browser on startup when prompted
  - Add better error handling to the startup-docker.sh script
  - Automatically detect and build binaries if they're missing when running locally using local build

  - When using the local build (i.e. not building in Docker), mount the binary dirs as volumes instead of copying them from context

This is more efficient and makes building images much faster. This also means images don't need to be re-built every time the binaries are updated locally - they just need to be restarted.

  - Make it easier to build in Docker by adding a new option to the startup-docker.sh script

  - Initialise services in the correct order in the startup-docker.sh script

Different PN services need to wait for each other's metadata. This causes a problem where all apps may be "initialised" according to Docker but journeys would fail because not all services would have had a chance to fetch the metadata they need. Unlike Kubernetes, Docker doesn't wait for the health checks to pass and it's up to the user to initialise services in the correct order. This is done using the wait-for-it.sh script

  - Add ability to build VSP in Docker

Added a new local image for the VSP to bring it in line with the other Proxy Node apps.