EID-1960 Apply connector node config to the correct pipeline job
Ensure that connector node config changes in the `verify-eidas-config` repo are applied only to the correct environment.

i.e. a change in `verify-eidas-config/proxy-node/integration` will trigger an integration deployment, but will not trigger a new build or a production deploy.

Any change in `verify-eidas-config/proxy-node/test` will trigger a new build, and subsequently a deploy to integration and production.