EID-1919 Configure Gateway to use QWAC
add `qwacCertificate.enabled: false` to `values.yaml` so that the default is to not use QWAC

add conditional sections to `gateway-ingress-gateway.yaml` around the `tls` and `hosts` section so that when `qwacCertificate.enabled` is true, `tls` and `hosts` are handled differently.

add `--set "global.qwacCertificate=true"` to inline bash used in `render-manifests` in `prod/depoly-pipeline.yaml` so that production uses the QWAC

cross fingers and hope for the best.