BAU Use parallelStream in MetadataResolverService constructor
The MetadataResolverService needs to create a MetadataResolver for each configured country.  When it does this in series then any faulty countries must time out before the next country can be resolved.  As the resolvers are created during a stream operation we can replace `stream()` with `parallelStream()`.  Using the integration configuration it can reduce the time taken to create the resolvers from ~42s to ~12s. Of the 12s, at least 10 are the time it takes for countries to timeout.  Reducing the timeout numbers in `getClient()` would further speed up the startup.

Reduce CONNECT_TIMEOUT and READ_TIMEOUT in `getClient()` from 10_000 to 5_000.