BAU Add logging filters to Metatron
`MetatronResource` was annotated with `@IngressEgressLogging` but the appropriate filters were not registered so nothing gets logged.  I'll admit, I'm not entirely sure what will be recorded when the logging does commence as Metatron is pretty lightweight in terms of it's calls so I'm not sure what information will be available to log.