WIP Killswitch
I don't really expect this to work but it might just have the parts it needs to do what we want it to.

Ideally we'd be able to see the lock file writing happen (lock/unlock) and then see if that happens when the killswitch is triggered.  Then we could add in a check for the lockfile to the deploy stuff.