EID-2020: Fix Netherlands smoke test
Although running locally with'-c' flag  works fine. concourse is complaining
'Element <a href="#"> is not clickable at point (537.75,377.75) because another element <span> obscures it '

So running without the '-c' flag gives the same error locally. This change has fixed this error lcoally, but the final step seems to fail.
But the last step seems to pass in all the tests on concourse so we can assume that it will work.