EID-2020: Attempt 2 fix netherlands smoke test
Currentely the smoke test is being flakey.
The click_link is what ive discovered to cause the issues.
Netherlands anchor tag is set up like this href="#", i believe the clink_link is having issues with this.
So being more specific looking for the tag to click on might solve this.