BAU: Fix Italian proxy node test
The Italian proxy node test is _very_ flaky. We often see the test
failing at the Italians connector-node. There are a series of pages
which are auto-submitted with JS. One particular page's JS is fired by a
`load` event on the window. There may be a race condition as a lot of
the time this doesn't happen with Selenium and the journey stalls.

This adds a conditional check to see if we've stalled. If so, we
manually call the JS required to submit the page.